--- agile-app/src/lib/carouselPhotos.ts
@@ -1,14 +1,0 @@
-export const carouselPhotos = [
-    "covered_deck.avif",
-    "RY_Kitchen.avif",
-    "RY_Truss.avif",
-    "AddOnR5.avif",
-    "AddOnRoom.avif",
-    "BY_Remodel.avif",
-    "BY_InsideRemodel.avif",
-    "D_house.avif",
-    "D_Garage.avif",
-    "Night_Deck.avif",
-    "Porch1.avif",
-    "Porch2_1.avif",
-  ];
