--- agile-app/src/lib/familyPhotos.ts
@@ -1,8 +1,0 @@
-export const familyPhotos = [
-    "Family.avif",
-    "AshAndDad.avif",
-    "DadJeffAndPapa.avif",
-    "Family2.avif",
-    "lauren_grad.avif",
-    "WillGrad.avif"
-];
