--- agile-app/src/lib/galleryPhotos.ts
@@ -1,51 +1,0 @@
-export const galleryPhotos = [
-  "AddOnR1.avif",
-  "AddOnR2.avif",
-  "AddOnR3.avif",
-  "AddOnR4.avif",
-  "AddOnR5.avif",
-  "AddOnRoom.avif",
-  "BY_InsideRemodel.avif",
-  "BY_Remodel.avif",
-  "covered_deck_2.avif",
-  "covered_deck.avif",
-  "D_Bathroom.avif",
-  "D_Garage.avif",
-  "D_house.avif",
-  "dvile_inside.avif",
-  "L_Shower.avif",
-  "L_Sink.avif",
-  "Night_Deck.avif",
-  "Night_deck2.avif",
-  "Perg1.avif",
-  "Perg2.avif",
-  "Perg3.avif",
-  "Perg4.avif",
-  "Porch1.avif",
-  "Porch2_1.avif",
-  "Porch2.avif",
-  "PorchWalk3.avif",
-  "Red_Bath.avif",
-  "Red_Shower.avif",
-  "RY_Bath.avif",
-  "RY_Bath2.avif",
-  "RY_Bed2.avif",
-  "RY_Bed3.avif",
-  "RY_Bed4.avif",
-  "RY_Bedroom.avif",
-  "RY_Cab.avif",
-  "RY_Counter.avif",
-  "RY_CP.avif",
-  "RY_CP2.avif",
-  "RY_FP.avif",
-  "RY_house.avif",
-  "RY_House2.avif",
-  "RY_Kitch.avif",
-  "RY_Kitch2.avif",
-  "RY_Kitchen.avif",
-  "RY_Living.avif",
-  "RY_Living2.avif",
-  "RY_Truss.avif",
-  "WDeck.avif",
-  "WDeck2.avif"
-];
