+++ agile-app/src/services/familyPhotos.ts
@@ -1,0 +1,8 @@
+export const familyPhotos = [
+    "Family.avif",
+    "AshAndDad.avif",
+    "DadJeffAndPapa.avif",
+    "Family2.avif",
+    "lauren_grad.avif",
+    "WillGrad.avif"
+];
